+++ src/components/tmc2209_arduino/src/TMC2209_impl.h
@@ -36,7 +36,7 @@ void tmc2209_c_init(stepper_motor_instances_t instance);
 
 void tmc2209_c_enable(stepper_motor_instances_t instance);
 
-void tmc2209_c_set_speed(stepper_motor_instances_t instance, int16_t speed);
+void tmc2209_c_set_speed(stepper_motor_instances_t instance, int32_t speed);
 
 void tmc2209_c_disable(stepper_motor_instances_t instance);
